+++ torchci/components/benchmark_v3/configs/configurations.tsx
@@ -177,7 +177,7 @@ export const BENCHMARK_CATEGORIES: BenchmarkCategoryGroup[] = [
       },
       {
         name: "PyTorch x vLLM Benchmark",
-        route: `/benchmark/v3/dashboard/${PYTORCH_X_VLLM_BENCHMARK_ID}`,
+        route: `/benchmark/v3/aggregate/${PYTORCH_X_VLLM_AGGREGATE_BENCHMARK_ID}`,
         info: "PyTorch x vLLM nightly benchmark using [vLLM pinned commit](https://github.com/pytorch/pytorch/blob/main/.github/ci_commit_pins/vllm.txt). Powered by [vllm-benchmark workflow](https://github.com/pytorch/pytorch/blob/main/.github/workflows/vllm-benchmark.yml) + [the benchmark configs](https://github.com/pytorch/pytorch-integration-testing/tree/main/vllm-benchmarks/benchmarks)",
         description: "Pytorch x vLLM nightly benchmark on PyTorch",
         actions: [
@@ -187,8 +187,8 @@ export const BENCHMARK_CATEGORIES: BenchmarkCategoryGroup[] = [
             href: `/benchmark/regression/reports/${PYTORCH_X_VLLM_BENCHMARK_ID}`,
           },
           {
-            label: "Aggregated Dashboard",
-            href: `/benchmark/v3/aggregate/${PYTORCH_X_VLLM_AGGREGATE_BENCHMARK_ID}`,
+            label: "Detail Dashboard",
+            href: `/benchmark/v3/dashboard/${PYTORCH_X_VLLM_BENCHMARK_ID}`,
           },
         ],
       },
