+++ torchci/lib/benchmark/api_helper/backend/dataFetchers/fetchers.ts
@@ -4,8 +4,6 @@ import {
   PytorchHelionDataFetcher,
   PytorchOperatorMicroBenchmarkDataFetcher,
   VllmBenchmarkDataFetcher,
-  VllmXPytorchBenchmarkAggregatedDataFetcher,
-  VllmXPytorchBenchmarkDataFetcher,
 } from "./queryBuilderUtils/benchmarkDataQueryBuilder";
 import {
   BenchmarkListCommitQueryBuilder,
@@ -24,6 +22,10 @@ import {
   BenchmarkListCommitFetcher,
   BenchmarkMetadataFetcher,
 } from "./type";
+import {
+  VllmXPytorchBenchmarkAggregatedDataFetcher,
+  VllmXPytorchBenchmarkDataFetcher,
+} from "./queryBuilderUtils/vllmDataFetchers";
 
 // Register benchmark data fetchers, this is mainly used in get_benchmark_data api and get_time_series api
 const dataCtors: Record<string, new () => BenchmarkDataFetcher> = {
